# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Xeno CRM API",
      "version": "1.0.0",
      "description": "API for Xeno CRM platform, including audience management, campaigns, AI tools, and more."
    },
...
    "paths": {
...
      "/api/ai/suggest-messages": {
        "post": {
          "tags": ["AI"],
          "summary": "Suggests marketing messages based on campaign objectives.",
          "description": "Rate limit: 5 requests per minute per user (Upstash Redis). Exceeding this limit will redirect to the home page (/).",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuggestMessagesRequest" }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successfully suggested messages.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "suggestions": {
                        "type": "array",
                        "items": { "type": "string" }
                      }
                    }
                  }
                }
              }
            },
            "400": { "$ref": "#/components/responses/BadRequest" },
            "422": { "$ref": "#/components/responses/UnprocessableEntity" },
            "500": { "$ref": "#/components/responses/InternalServerError" },
            "503": { "$ref": "#/components/responses/ServiceUnavailable" }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "SuggestMessagesRequest": {
          "type": "object",
          "properties": {
            "objective": { "type": "string", "minLength": 10, "description": "Campaign objective." },
            "audienceDescription": { "type": "string", "nullable": true, "description": "Description of the target audience." },
            "tone": { "type": "string", "enum": ["neutral", "formal", "friendly", "playful", "urgent"], "default": "neutral", "nullable": true },
            "messageCount": { "type": "integer", "minimum": 1, "maximum": 5, "default": 3, "nullable": true }
          },
          "required": ["objective"]
        },
...
      },
...
      "responses": {
        "BadRequest": {
          "description": "Bad Request - Often due to validation errors.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string" },
                  "errors": { "type": "object", "additionalProperties": true, "nullable": true }
                }
              }
            }
          }
        },
...
        "UnprocessableEntity": {
          "description": "Unprocessable Entity - The server understands the content type of the request entity, and the syntax of the request entity is correct, but it was unable to process the contained instructions (e.g., AI could not process prompt).",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string" }
                }
              }
            }
          }
        },
        "InternalServerError": {
          "description": "Internal Server Error - An unexpected condition was encountered.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string" },
                  "error": { "type": "string", "nullable": true }
                }
              }
            }
          }
        },
        "ServiceUnavailable": {
          "description": "Service Unavailable - The server is currently unable to handle the request due to a temporary overload or maintenance (e.g., AI service not configured).",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string" }
                }
              }
            }
          }
        },
...
      }
    }
  }